MODSOURMAN-854: Code improvements.
