+++ mod-source-record-manager-server/src/main/java/org/folio/verticle/consumers/errorhandlers/payloadbuilders/MarcBibDiErrorPayloadBuilder.java
@@ -1,34 +1,20 @@
 package org.folio.verticle.consumers.errorhandlers.payloadbuilders;
 
 import io.vertx.core.Future;
-import io.vertx.core.json.Json;
-import io.vertx.core.json.JsonArray;
-import io.vertx.core.json.JsonObject;
-import org.apache.commons.lang3.StringUtils;
 import org.folio.DataImportEventPayload;
 import org.folio.dataimport.util.OkapiConnectionParams;
-import org.folio.rest.jaxrs.model.ParsedRecord;
 import org.folio.rest.jaxrs.model.Record;
 import org.folio.services.MappingRuleCache;
 import org.folio.services.entity.MappingRuleCacheKey;
-import org.folio.verticle.consumers.errorhandlers.RawMarcChunksErrorHandler;
-import org.folio.services.util.RecordConversionUtil;
-import org.folio.verticle.consumers.util.MarcImportEventsHandler;
+import org.folio.verticle.consumers.util.DiErrorBuilderUtil;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
 
-import java.util.HashMap;
-import java.util.Map;
-import java.util.Optional;
-
 import static org.folio.rest.jaxrs.model.Record.RecordType.MARC_BIB;
-import static org.folio.rest.jaxrs.model.DataImportEventTypes.DI_ERROR;
-import static org.folio.verticle.consumers.util.MarcImportEventsHandler.NO_TITLE_MESSAGE;
 
 @Component
 public class MarcBibDiErrorPayloadBuilder implements DiErrorPayloadBuilder {
-  private static final String FIELDS = "fields";
-  private MappingRuleCache mappingRuleCache;
+  private final MappingRuleCache mappingRuleCache;
 
   @Autowired
   public MarcBibDiErrorPayloadBuilder(MappingRuleCache mappingRuleCache) {
@@ -48,65 +34,12 @@ public class MarcBibDiErrorPayloadBuilder implements DiErrorPayloadBuilder {
 
     return mappingRuleCache.get(new MappingRuleCacheKey(okapiParams.getTenantId(), record.getRecordType()))
       .compose(rulesOptional -> {
-        DataImportEventPayload diErrorPayload = prepareDiErrorEventPayload(throwable, okapiParams, jobExecutionId, record);
+        DataImportEventPayload diErrorPayload = DiErrorBuilderUtil.prepareDiErrorEventPayload(throwable, okapiParams, jobExecutionId, record);
         if (rulesOptional.isPresent()) {
-          String recordContent = getReducedRecordContentOnlyWithTitle(rulesOptional.get(), record);
-          return Future.succeededFuture(makeLightweightPayload(record, recordContent, diErrorPayload));
+          String recordContent = DiErrorBuilderUtil.getReducedRecordContentOnlyWithTitle(rulesOptional.get(), record);
+          return Future.succeededFuture(DiErrorBuilderUtil.makeLightweightPayload(record, recordContent, diErrorPayload));
         }
-        return Future.succeededFuture(makeLightweightPayload(record, null, diErrorPayload));
+        return Future.succeededFuture(DiErrorBuilderUtil.makeLightweightPayload(record, null, diErrorPayload));
       });
-  }
-
-  private DataImportEventPayload prepareDiErrorEventPayload(Throwable throwable,
-                                                            OkapiConnectionParams okapiParams,
-                                                            String jobExecutionId, Record record) {
-    HashMap<String, String> context = new HashMap<>();
-    context.put(RawMarcChunksErrorHandler.ERROR_KEY, throwable.getMessage());
-    context.put(RecordConversionUtil.getEntityType(record).value(), Json.encode(record));
-
-    return new DataImportEventPayload()
-      .withEventType(DI_ERROR.value())
-      .withJobExecutionId(jobExecutionId)
-      .withOkapiUrl(okapiParams.getOkapiUrl())
-      .withTenant(okapiParams.getTenantId())
-      .withToken(okapiParams.getToken())
-      .withContext(context);
-
-  }
-
-  private String getReducedRecordContentOnlyWithTitle(JsonObject mappingRules, Record record) {
-    Optional<String> titleFieldOptional = MarcImportEventsHandler.getTitleFieldTagByInstanceFieldPath(mappingRules);
-    if (titleFieldOptional.isPresent()) {
-      String titleFieldTag = titleFieldOptional.get();
-
-      ParsedRecord parsedRecord = record.getParsedRecord();
-      if (parsedRecord == null) {
-        return new JsonObject()
-          .put(FIELDS, new JsonArray()
-            .add(new JsonObject()
-              .put(titleFieldTag, NO_TITLE_MESSAGE))).encode();
-      }
-      JsonObject parsedContent = new JsonObject(parsedRecord.getContent().toString());
-      var fields = parsedContent.getJsonArray(FIELDS).getList();
-      for (Object elem: fields) {
-        Object titleField = ((Map) elem).get(titleFieldTag);
-        if (titleField != null) {
-          return new JsonObject()
-            .put(FIELDS, new JsonArray()
-              .add(new JsonObject()
-                .put(titleFieldTag, titleField))).encode();
-        }
-
-      }
-    }
-    return null;
-  }
-
-  private DataImportEventPayload makeLightweightPayload(Record record, String newRecordContent, DataImportEventPayload payload) {
-    record.setParsedRecord(StringUtils.isBlank(newRecordContent) ? null : new ParsedRecord().withContent(newRecordContent));
-    record.setRawRecord(null);
-    payload.setProfileSnapshot(null);
-    payload.getContext().put(RecordConversionUtil.getEntityType(record).value(), Json.encode(record));
-    return payload;
   }
 }
